Add TileInfo resource
Instead of setting a floor tile and a wall tile in the Generator, you pass in a TileInfo resource. This resource contains the position of the tile in the TileSet, and if it is a single cell or a terrain.
Now, the Generate Walls (renamed to Generate Borders) modifier takes in a TileInfo aswell to determine which tile to use as a border.